Bugfix OverviewActivity + event title rediction
Saturday was removed from options in OverviewActivity because no
schedule will ever be in the application on saturday.

Changes (like reduction of the length of event's title and schrinking
the size of font in which titles are showed) were made in order to make
constant space for one event in Overview even for smaller devices. It's
were made because the animation of switching between days, looks much
better this way.